requirements.txt
azure-cognitiveservices-speech==1.23.0
PyAudio==0.2.12
SpeechRecognition==3.8.1


